backend/requirements.txt
pdf2image>=1.16.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
Pillow>=9.0.0
numpy>=1.24.0
tqdm>=4.60.0
